# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: BigIn Asset API
  description: >-
    API quản lý tài sản BigIn Asset. Bao gồm xác thực (F01), quản lý tài sản (F02),
    phiếu mượn (F03), duyệt & giữ chỗ (F04), bàn giao/hoàn trả/lịch sử mượn (F05),
    sự cố & sửa chữa (F06), thông báo (F07) và quản trị người dùng/phân quyền (F08).

    Quy ước chung:
    - Envelope thành công: `{ "data": ... }`.
    - Envelope lỗi: `{ "error": string, "details"?: object }`.
    - Danh sách: `page=1`, `pageSize=20` (tối đa 100); page DTO `{ items, page, pageSize, total }`.
    - Danh sách sắp xếp mới nhất trước; Review Queue sắp xếp cũ nhất trước.
    - `expectedReturnDate` chỉ ngày `YYYY-MM-DD` theo múi giờ `Asia/Ho_Chi_Minh`, phải từ hôm nay trở đi.
    - Authorization theo effective permission code (không kiểm tra tên role).
    - Lỗi dự kiến: `400` input sai, `401` chưa xác thực, `403` thiếu quyền, `404` không tìm thấy, `409` xung đột trạng thái/dữ liệu.
  version: 1.0.0
  contact:
    name: BigIn SE Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /borrow-request-details/handover-queue:
    get:
      tags:
        - F05 - Handover, Return & History
      summary: Hàng đợi bàn giao
      description: >-
        Yêu cầu permission `asset.checkout`. Chỉ trả về detail `APPROVED` có asset
        `RESERVED` và chưa có borrow history; không yêu cầu `borrow_request.view_all`
        hoặc `borrow_history.view_all`. Sắp xếp request cũ trước, sau đó `detailId`
        tăng dần.
      operationId: listHandoverQueue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Trang hàng đợi bàn giao
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HandoverQueuePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /borrow-request-details/review-queue:
    get:
      tags:
        - F04 - Approval & Reservation
      summary: Hàng đợi duyệt
      description: >-
        Trang toàn công ty chứa các phiếu có ít nhất một detail `PENDING`, sắp xếp cũ nhất trước.
      operationId: reviewQueue
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: approvalStatus
          in: query
          description: Lọc theo trạng thái duyệt của detail (mặc định PENDING)
          schema:
            $ref: '#/components/schemas/BorrowDetailStatus'
      responses:
        '200':
          description: Trang hàng đợi duyệt
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BorrowRequestPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /borrow-histories/return-queue:
    get:
      tags:
        - F05 - Handover, Return & History
      summary: Hàng đợi nhận trả
      description: >-
        Yêu cầu permission `asset.checkin`. Chỉ trả về borrow history chưa có
        `return_date`; không yêu cầu `borrow_request.view_all` hoặc
        `borrow_history.view_all`. Sắp xếp theo expected return date gần nhất/quá
        hạn trước, rồi `borrowedAt` cũ trước và `historyId` tăng dần.
      operationId: listReturnQueue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Trang hàng đợi nhận trả
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BorrowHistoryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      description: Số trang, bắt đầu từ 1
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Số bản ghi mỗi trang, tối đa 100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Chưa xác thực hoặc token không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Thiếu quyền thực hiện thao tác
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    MediaEvidence:
      type: object
      required: [mediaId, mimeType, sizeBytes, uploadedAt, publicUrl]
      properties:
        mediaId: { type: integer }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        uploadedAt: { type: string, format: date-time }
        publicUrl: { type: string, format: uri }
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
...
    AssetStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - BORROWED
        - DAMAGED
        - IN_REPAIR
        - RETIRED
...
    AssetSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
...
    BorrowRequestStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - PARTIALLY_APPROVED
        - COMPLETED
        - CANCELLED
    BorrowDetailStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
...
    BorrowRequest:
      type: object
      properties:
        id:
          type: integer
        requester:
          type: object
          properties:
            id:
              type: integer
            userCode:
              type: string
              pattern: '^BI[0-9]{5,}$'
            name:
              type: string
            avatarUrl:
              type: string
              nullable: true
          required:
            - id
            - userCode
            - name
        status:
          $ref: '#/components/schemas/BorrowRequestStatus'
        note:
          type: string
          maxLength: 300
        createdAt:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              asset:
                type: object
                properties:
                  id:
                    type: integer
                  serialNumber:
                    type: string
                    nullable: true
                  qrCode:
                    type: string
                  imageUrl:
                    type: string
                    nullable: true
                  status:
                    $ref: '#/components/schemas/AssetStatus'
                  model:
                    $ref: '#/components/schemas/AssetSummary'
                required:
                  - id
                  - qrCode
                  - status
                  - model
              expectedReturnDate:
                type: string
                format: date
              approvalStatus:
                $ref: '#/components/schemas/BorrowDetailStatus'
              approvedBy:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                required:
                  - id
                  - name
                nullable: true
              approvedAt:
                type: string
                format: date-time
                nullable: true
              rejectionReason:
                type: string
                maxLength: 300
                nullable: true
            required:
              - id
              - asset
              - expectedReturnDate
              - approvalStatus
          required:
            - id
            - requester
            - status
            - createdAt
            - details
      required:
        - id
        - requester
        - status
        - createdAt
        - details
    BorrowRequestListItem:
      type: object
      properties:
        id:
          type: integer
        status:
          $ref: '#/components/schemas/BorrowRequestStatus'
        createdAt:
          type: string
          format: date-time
        detailCount:
          type: integer
        pendingCount:
          type: integer
        approvedCount:
          type: integer
        rejectedCount:
          type: integer
      required:
        - id
        - status
        - createdAt
        - detailCount
        - pendingCount
        - approvedCount
        - rejectedCount
    BorrowRequestPage:
      type: object
      properties:
        items:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/BorrowRequestListItem'
              - $ref: '#/components/schemas/BorrowRequest'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required:
        - items
        - page
        - pageSize
        - total
...
    HandoverQueueItem:
      type: object
      properties:
        detailId:
          type: integer
        requestId:
          type: integer
        requestCreatedAt:
          type: string
          format: date-time
        requester:
          type: object
          properties:
            id:
              type: integer
            userCode:
              type: string
              pattern: '^BI[0-9]{5,}$'
            name:
              type: string
            email:
              type: string
              format: email
            avatarUrl:
              type: string
              nullable: true
            department:
              type: object
              nullable: true
              properties:
                id:
                  type: integer
                name:
                  type: string
              required:
                - id
                - name
          required:
            - id
            - userCode
            - name
            - email
            - avatarUrl
            - department
        asset:
          type: object
          properties:
            id:
              type: integer
            serialNumber:
              type: string
              nullable: true
            qrCode:
              type: string
            imageUrl:
              type: string
              nullable: true
            status:
              $ref: '#/components/schemas/AssetStatus'
            model:
              $ref: '#/components/schemas/AssetSummary'
          required:
            - id
            - serialNumber
            - qrCode
            - imageUrl
            - status
            - model
        expectedReturnDate:
          type: string
          format: date
        approvedBy:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        approvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - detailId
        - requestId
        - requestCreatedAt
        - requester
        - asset
        - expectedReturnDate
        - approvedBy
        - approvedAt
    HandoverQueuePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HandoverQueueItem'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required:
        - items
        - page
        - pageSize
        - total
...
    MediaEvidenceList:
      type: array
      items:
        $ref: '#/components/schemas/MediaEvidence'
...
    BorrowHistory:
      type: object
      properties:
        id:
          type: integer
        detailId:
          type: integer
        asset:
          type: object
          properties:
            id:
              type: integer
            serialNumber:
              type: string
              nullable: true
            qrCode:
              type: string
            imageUrl:
              type: string
              nullable: true
            status:
              $ref: '#/components/schemas/AssetStatus'
            model:
              $ref: '#/components/schemas/AssetSummary'
          required:
            - id
            - qrCode
            - status
            - model
        borrower:
          type: object
          properties:
            id:
              type: integer
            userCode:
              type: string
              pattern: '^BI[0-9]{5,}$'
            name:
              type: string
            avatarUrl:
              type: string
              nullable: true
          required:
            - id
            - userCode
            - name
        expectedReturnDate:
          type: string
          format: date
        handedOverBy:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
          nullable: true
        borrowedAt:
          type: string
          format: date-time
        receivedBy:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
          nullable: true
        returnedAt:
          type: string
          format: date-time
          nullable: true
        returnCondition:
          type: string
          nullable: true
        handoverEvidence:
          $ref: '#/components/schemas/MediaEvidenceList'
        returnEvidence:
          $ref: '#/components/schemas/MediaEvidenceList'
      required:
        - id
        - detailId
        - asset
        - borrower
        - expectedReturnDate
        - borrowedAt
        - handoverEvidence
        - returnEvidence
...
    BorrowHistoryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BorrowHistory'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required:
        - items
        - page
        - pageSize
        - total
...
security:
  - bearerAuth: []